+++ src/Resources/Resources/ChangeLog.md
@@ -19,7 +19,10 @@
 -->
 
 ## Upcoming Release
+
+## Version 6.6.0
 * Fixed an issue when running the `New-AzManagementGroup` command where it tried to cast an async operation as a Management Group. [#21000]
+* Fixed an issue when running `Get-AzResourceGroup -Name 'some_name'`, it ignores '-Location' if specified[#21225]
 
 ## Version 6.5.3
 * Updated behavior of Get-AzPolicyDefinition which previously returned all definitions when -Id was provided with a nonexistent policy definition id. Fixed to correctly throw a 404 exception in this case.