+++ src/Storage/Storage.Management/ChangeLog.md
@@ -18,6 +18,26 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added warning messages for an upcoming cmdlet breaking change 
+    - `New-AzStorageAccount`
+    - `Set-AzStorageAccount`
+
+## Version 5.5.0
+* Supported create storage account with DnsEndpointType
+    - `New-AzStorageAccount`
+* Fixed file cmdlets potential context issue when the current context doesn't match with the credential of input Azure File object
+    - `Close-AzStorageFileHandle`
+    - `Get-AzStorageFile`
+    - `Get-AzStorageFileContent`
+    - `Get-AzStorageFileHandle`
+    - `New-AzStorageDirectory`
+    - `New-AzStorageFileSASToken`
+    - `Remove-AzStorageDirectory`
+    - `Remove-AzStorageFile`
+    - `Remove-AzStorageShare`
+    - `Set-AzStorageFileContent`
+    - `Set-AzStorageShareQuota`
+    - `Start-AzStorageFileCopy`
 
 ## Version 5.4.2
 * Supported new Blob Tier Cold